Move origin validation function to package/ident
