+++ components/core/src/lib.rs
@@ -73,7 +73,6 @@ pub mod os;
 pub mod output;
 pub mod event;
 pub mod channel;
-pub mod origin;
 
 use std::path::PathBuf;
